Archdep: Make archdep_stat() properly return `len`.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38150 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/shared/archdep_stat.c
@@ -49,7 +49,7 @@
  *
  * \return  0 on success, -1 on failure
  */
-int archdep_stat(const char *path, unsigned int *len, unsigned int *isdir)
+int archdep_stat(const char *path, size_t *len, unsigned int *isdir)
 {
     struct stat statbuf;
