+++ vice/src/arch/shared/archdep_stat.h
@@ -27,6 +27,8 @@
 #ifndef ARCHDEP_STAT_H
 #define ARCHDEP_STAT_H
 
-int archdep_stat(const char *filename, unsigned int *len, unsigned int *isdir);
+#include <stddef.h>
+
+int archdep_stat(const char *filename, size_t *len, unsigned int *isdir);
 
 #endif
